+++ enterprise/api.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+"""
+Helper functions for enterprise app.
+"""
+
+from enterprise.models import EnterpriseCustomerBrandingConfiguration
+
+
+def get_enterprise_branding_info(provider_id=None):
+    """
+    Return the EnterpriseCustomer branding information.
+    """
+    return EnterpriseCustomerBrandingConfiguration.objects.filter(
+        enterprise_customer__enterprise_customer_identity_provider__provider_id=provider_id
+    ).first()